Initial commit: n-gram generator
This is the initial commit of the n-gram generation program. The plan is
to use it to generate custom code books for a fixed dictionary based
compressor, like the smaz (<https://github.com/antirez/smaz>)
compressor. It could be used for other purposes, which the Wikipedia
page of n-grams hints at (see <https://en.wikipedia.org/wiki/N-gram>).
The n-gram program is pretty much fully formed, there are a few things
to do (that need doing) and some nice-to-have things.

Need to do:
- Print out ranges of n-grams, one per line, with more formatting
options
- Add more optional sets

Nice to have:
- Being able to split input into tokens based on a regular expression